Skip NULL VariantType test with variant_types due to Spark codegen bug
Spark's codegen calls getValue() on null VariantVal when serializing NULL
VariantType values with variant_types option, causing NullPointerException.
This is a known Spark issue that needs to be fixed upstream.

The test is marked as ignore with a TODO to re-enable once Spark fixes
NULL VariantType serialization with variant_types.